feat(refactor): refactor app

+++ lib/shop/data/model/response/cart/cart_update_response_model.dart
@@ -3,10 +3,10 @@ import 'package:shop_app_clean_architecture/shop/data/model/json_converter.dart'
 class CartUpdateResponseModel
     implements JsonConverter<CartUpdateResponseModel> {
   const CartUpdateResponseModel({
-     this.total,
+    this.total,
   });
 
-  final double? total;
+  final int? total;
 
   @override
   CartUpdateResponseModel fromJson(Map<String, dynamic>? jsonMap) {
